+++ trunk/test_app/objects/monitor.c
@@ -10,7 +10,7 @@
 
 using namespace std;
 
-int calculate_monitor( obj_loc_t obj_loc, int id )
+int calculate_monitor( obj_loc_t obj_loc, long id )
 {
     T4_TEST( tier2_p_int, 14965 );
     T4_TEST( tier1s_int_out(0), 15003 );
